I have added Agecalculator project using html,css,js
The Age Calculator project is a web application built with HTML, CSS, and JavaScript that enables users to input their birthdate and calculates their current age, displaying it instantly in a visually appealing and user-friendly interface. It leverages client-side scripting to compute the age based on the input date, making it a simple and interactive tool.